requirements.txt
flask
pymongo
cloudinary
dotenv
deepface
dlib
python-telegram-bot==13.13